:sparkles: navigationbar상단으로 올리는 기능 추가
